simplification: omit action from pending queue
The existing code was storing the source action (in addition to the resulting
state) in the pending queue. That wasn't necessary since the action is
regenerated in `path_to`.

The resulting code is a bit simpler, but note that benchmarks did not improve
as the pending queue does not account for a significant part of check time.