loosen type constraints
This commit moves some type constraints from traits and structs to functions.
Doing so allows a wider range of valid cases to compile. For example, an actor
might persist a `SocketAddr` in its state for real-world execution, but
`SocketAddr` does not currently implement `Ord`, so a model would fail to type
check. This commit allows a parameterized actor to work in both cases.