move examples into different applications
This approach will result in more applications (and corresponding boilerplate),
but it keeps all the code for a particular state machine example in a single
file, which I suspect is more convenient for someone learning the library.

It also avoids duplicate test runs, which the old implementation would do
because each example application imported the entire `state_machines` module.