introduce `StateMachine::Action` and `QuickMachine`
In prior releases, an action was simply a string describing a step. Now
it is a general value, which is useful for debugging as the library can
provide detailed invariant violation "paths." For example, an invariant
violated by an actor system now features a step-by-step description of
the messages and individual actor state changes rather than simply
indicating a sequence of broader system states along with a string
roughly describing each step.

Model checking is slightly faster  in many cases as state machines no
longer return a pair containing both the new state and the description.
Instead the "description" (action) is returned, which gets converted
into the new state. On a restored laptop for example, I observed the
following changes in model check times:

- 2pc 8: 22 --> 19 sec
- paxos 4: 13 --> 13 sec
- wor 6: 24 --> 21 sec

The commit also introduces a `QuickMachine` structure, which eliminates
some boilerplate to make examples a bit more succinct.