allow de/serialization overrides
This commit introduces a layer of indirection for message
de/serialization between an actor state machine and the external world.
This is useful when composing actor systems.

As an example, if one actor handles messages `X` and `Y`, while another
handles messages `Y` and `Z`, we might want to be able to compose the
systems based on the message `Y`.  This can be done at runtime as the
messages will deserialize appropriately if the structures match, but
model checking would require the introduction of a wrapper message type.
A gateway allows you to remove the wrapper type at runtime, which serves
two purposes: messages are more succinct, and an existing system can be
leveraged as a dependency of a new system without recompiling to
introduce a wrapper type.