introduce `Property` and `Model`
A `Property` is a named predicate that generalizes the notion of an
invariant.  Properties can be of category "sometimes" or "always."  The
checker looks for examples of "sometimes" properties and counterexamples
of "always" properties.

A `Model` combines a `StateMachine` with a collection of properties. More
model checking options will be added to the `Model` type in upcoming
commits.

This commit also introduces improved performance by replacing the
`Worker` type with a simplification that reduces copying. Performance on
an older laptop changed as follows:

- 2pc 8   @ 1,745,408 states: 19 -> 11 sec
- paxos 4 @   479,167 states: 14 -> 10 sec
- wor 6   @ 1,576,960 states: 22 -> 11 sec

Some additional changes are also included:

- Revise the crate description.
- Refactor to reduce some boilerplate.
- Introduce more tests.
- Add `explore` mode to the `wor` example.
- Replace the `fxhash` dependency with `ahash` and `dashmap`.
- Move `num_cpus` into a dev dependency.