introduce timers
These allow an actor to schedule a timeout, enabling a wider class of use cases
such as message resends and leader re-election.

Performance impact is relatively small for actors that do not use timers:

paxos 5: 22  --> 23 sec
wor   7: 114 --> 116 sec

That said, the current implementation is completely unoptimized for cases where
timers are used, and there is room for improvement. For example, the
`Vec<bool>` can be replaced with a `BitVec`.