fix: panic when setting timer in `on_start`
The actor system model has an optimization that only allocates the timer flags
if timers are used, and the existing implementation could infer the wrong
number of actors if called from `on_start`.

This commit also introduces an `Out::is_no_op` helper and adds `Hash` to the
`Actor::State` type constraint.