introduce `System::History`
This feature allows a system model to capture additional world state
outside that of the actors. It should be useful for checking properties
where history is relevant, such as linearizability.

This commit also simplifies `SystemState` and summarizes some key
features in the `README.md` file.

Model checking performance is virtually identical.