introduce linearizable `single-copy-register` example
To facilitate linearizability testing, the commit amends `register` to
capture request IDs, with some other minor changes. The `paxos` example
had to be amended accordingly, and the latest implementation has a
differing state space.

This commit also removes the `wor` (write-once register) example as it
it is not sufficiently different from `single-copy-register` to be worth
including.