introduce `DfsChecker`
This checker performs a depth-first search and uses much less memory.

See Graydon's suggestion for this along with DPOR here:
https://github.com/stateright/stateright/issues/7

One difference is that the visited state is unbounded, but as mentioned
in the commit summary the actual memory usage is very small in
comparison to BFS memory usage.

Temporary limitations that future commits will address:

- Omits tests.
- Omits support for `eventually` properties.
- Omits support for multithreaded checking.