ui: show recently visited path
I recently spent over half an hour debugging an issue wherein a simple
model was very slow. The problem turned out to be that the
linearizability tester was recording redeliveries of a particular
message without bound. That investigation would have been much faster
had Stateright Explorer shown a recently visited path, which is the
motivation for this feature.

This is a breaking change as the UI must now have access to a
`CheckerBuilder` (rather than a `Checker`) so that it can register a
visitor.