do not panic on invalid histories
Prior to this change `RegisterTestSystem` would panic if it encountered
an invalid history. This allowed it to provide more useful information
than a simple boolean, but stopping the checker wasn't the right
solution as invalid histories arise in practice (for instance due to
message redelivery).

A later commit can explore the idea of returning a failure description.