record `RegisterTestSystem` history at clients not service
This is a better choice as it can identify linearizability violations
(for instance) caused by request redelivery to the server and response
redelivery to the client. Measuring at the server in contrast treated
request redelivery as a new invocation and omitted response redelivery.

The ./bench.sh baseline for actor examples is different as this commit
changes the state space. The Paxos benchmark in particular requires far
more clients to reach a multi-second check time on my laptop.

    == 2pc check 9 ==
    Done. generated=10340352, sec=13
    Done. generated=10340352, sec=14
    Done. generated=10340352, sec=16
    == paxos check 15 ==
    Done. generated=8152321, sec=25
    Done. generated=8152321, sec=25
    Done. generated=8152321, sec=25
    == single-copy-register check 4 ==
    Done. generated=790072, sec=8
    Done. generated=790072, sec=8
    Done. generated=790072, sec=8
    == linearizable-register check 4 ==
    Done. generated=3357071, sec=24
    Done. generated=3357071, sec=24
    Done. generated=3357071, sec=24