do not record history for ignored incoming messages
This change ensures consistency testers do not flag consistency
violations for redeliveries if the client ignores them (e.g. due to
implemented redelivery protection, such as an ordered reliable link).

This commit also amends `RegisterActor::Client` to properly handle
redeliveries. Furthermore, the first client sends a second `Put` before
a `Get`, which enables the checker to identify linearizability
violations even for single-client systems when the service does not
properly handle redelivery of the earlier `Put` after the second `Put`.

A note for future improvement: there are scenarios where an actor
responds to redelivered messages without treating them as operation
returns. For example, it is common for a recipient to `Ack` redelivered
messages so that the sender stops sending them. This would happen if an
earlier `Ack` was lost/delayed, for instance. Improvements are pending
to handle that scenario as well.

Once again the ./bench.sh baseline for actor examples is different
due to the revised state space.

    == 2pc check 9 ==
    Done. generated=10340352, sec=14
    Done. generated=10340352, sec=16
    Done. generated=10340352, sec=16
    == paxos check 15 ==
    Done. generated=8152321, sec=26
    Done. generated=8152321, sec=26
    Done. generated=8152321, sec=27
    == single-copy-register check 4 ==
    Done. generated=400233, sec=5
    Done. generated=400233, sec=5
    Done. generated=400233, sec=5
    == linearizable-register check 4 ==
    Done. generated=3357071, sec=25
    Done. generated=3357071, sec=24
    Done. generated=3357071, sec=26