optimization: `Vec` and `BTreeSet` for `2pc` example
This commit roughly doubles the speed of the `2pc` example by taking
advantage of the fact that a resource manager can simply be identified
by an integer in `0..N`. That enables any map/set with an RM key to be
replaced with a `Vec`.

The commit also replaces `HashableHashSet` with `BTreeSet` for the
message set, which improves performance slightly more.

Before:

    == 2pc check 9 ==
    Done. generated=10340352, sec=16
    Done. generated=10340352, sec=18
    Done. generated=10340352, sec=17

After:

    == 2pc check 9 ==
    Done. generated=10340352, sec=8
    Done. generated=10340352, sec=9
    Done. generated=10340352, sec=9

Since the default `2pc` benchmark now takes less than 10 seconds on my
laptop, this commit also bumps up the default to 10 resource managers:

    == 2pc check 10 ==
    Done. generated=61515776, sec=61
    Done. generated=61515776, sec=62
    Done. generated=61515776, sec=63

A future optimization could investigate using a bit vector for
`tm_prepared`.