correct linearizable register example
In the earlier implementation, a server coordinating a request would
overwrite its own sequencer and value when completing phase 1 even if
the quorum had a smaller sequencer.

Triggering the bug requires 3 servers, so that example now uses 3
servers for checking and exploring. Doing so dramatically increases the
state space, so `./bench.sh` now checks with 2 clients (which is still
sufficient to elicit the bug in the uncorrected implementation).

It also became clear that `within_boundary` was not needed for that
example, so it was removed as well.

As a related change, `RegisterActor::Client` now has a `put_count` field
so that the configuration can specify this value directly. Previously
the code relied upon a heuristic in which the first client would always
perform two puts while other clients would perform one, which isn't
needed in most cases. The commit alters the state spaces of the
benchmark as follows.

    == 2pc check 10 ==
    Done. generated=61515776, sec=60
    == paxos check 7 ==
    Done. generated=10957685, sec=21
    == single-copy-register check 4 ==
    Done. generated=400233, sec=5
    == linearizable-register check 2 ==
    Done. generated=41138199, sec=80