fix: `eq` and `partial_cmp` for `VectorClock`
The previous implementation of `partial_cmp` would only iterate over the
elements of the LHS (`self`) argument. That caused the following
assertion to fail.

    assert_eq!(
        Some(Less),
        VectorClock::from(vec![]).partial_cmp(&vec![1].into()));

It also had the compiler automatically implement `PartialEq`, which
cauesd the following assertions to fail.

    assert_eq!(
        VectorClock::from(vec![0]),
        VectorClock::from(vec![]));
    assert_eq!(
        VectorClock::from(vec![]),
        VectorClock::from(vec![0]));

This commit also implements `Eq` to enable vector clocks to be looked up
in hash maps/sets.