eliminate unwanted nondeterminism in linearizable-register
This unwanted nondeterminism would not manifest in the existing
implementation. Rather, I had intentionally introduced a bug locally
that caused the unexpected nondeterminism to manifest (as an error when
reconstructing a `Path` from fingerprints). I was surprised how easy it
was to trigger this mistake, so I've revised the code to make the
mistake harder to introduce.