explorer: quick nav along path
This commmit distinguishes between the "farthest state" for a path
and the currently selected state, which is characterized by an offset
from the farthest state.

Doing so enables quick navigation forward/backward along a path. This is
particulary quick with the newly provided keyboard navigation
(`down`/`up` or `j`/`k` keys for next/previous).

The commmit also changes the `ignored:` action prefix to an `(ignored)`
suffix and the `current state:` prefix to a `(state maches selected)`
suffix.