correct reindexing bug
Reindexing was buggy as prior to this commit it leveraged the same
mapping as that for rewrites, but in general reindexing and rewriting need
different maps (i.e. entries in the reindex mapping indicate "from where
did this element come," while entries in the rewrite mapping indicate
"to what value should this element become?").

Other notable changes:

- `Rewrite` now also accepts a type parameter indicating the rewritten type.
- `Reindex` is no longer needed and has been removed.

Follow-up work:

- `rewrite.rs` omits implementations for many types. It will be amended
   as needed.