correct symmetry reduction implementation
The previous implementation would enqueue the state from which the
representative state was derived, which did not necessarily match the
collected sequence of fingerprints for that state, resulting in the
occasional panic when reconstructing a path from those invalid
fingerprints. Those panics are no longer reproducible, but unfortunately
I was unable to prepare a regression test in the limited time that I
allocated for this fix.

This change increases the symmetry reduced state space. For instance,
with the 2PC example...

Before:

    $ cargo run --release --example 2pc check-sym 14
    ...
    Done. states=42458031, unique=2751591, sec=11

After:

    $ cargo run --release --example 2pc check-sym 14
    ...
    Done. states=322784853, unique=21523359, sec=69