fix: unintentional spin-wait loop in `spawn`
This addresses a bug introduced with named timers.

Even more importantly, the fact that the bug made it into a published
release highlights a gap in the test coverage for the real-world actor
runtime, which has always gotten less attention than the model checking
side of this library. I planned to add a regression test, but I believe
I have to first add support for an `Out::exit` command or similar and
don't want to block the fix on that.