Add Actor::on_random to model non-determinism
The actor::Out::choose_random method gives a way to model
non-deterministic choices by an Actor. It supplies a vec of choices with
type Actor::Random, which ActorModelState will then pick one. A string
key is also provided with the choose_random method, which can be used
to overwrite or remove previously registered random choices.

For explorer and DFS / BFS simulations, these random choices are modeled
as part of ActorModel::actions. For actor::spawn, the random choice is
is modeled as a timer with a duration between 0 - 10 seconds.

Because of the additions of the generic associated type Random and the
type parameters, this is a major semver change. Existing users of the
actor framework will need to add `type Random = ()`.

Because of the addition of string keys for random actions,
`ActorModelAction` also no longer implements `Copy`.