Fix LWW register inconsistency
In `merge` I left out updater_id component, that was just for testing to
deliberately trigger a failure. Another thing that I missed was that LWW
timestamps from the same client must be unique; that was fixed by
keeping track of the maximum used clock value and only generate events
with a newer clock.